Cleanup for_each_dupe
No actual functionality is changed. for_each_dupe() is moved into
duperemove.c with the rest of the extent finding code. The function callback
pointer is removed in favor of just calling walk_dupe_block directly. As a
result, we no longer need the dupe_walk_ctxt so that is unraveled. Naming of
filerec and block variables in these functions are changed to be more
consistent.

This makes reading through the extent finding code far easier.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>