dedupe_extent_list() fixes
We could exit the loop with a live context if there was exactly
MAX_DEDUPES_PER_IOCTL + 1 extents. The context gets cleaned up but that
means we might not dedupe that extent. Fix this by adding a previous extent
to the context and forcing a dedupe in that situation.

We also weren't bubbling up ENOMEM in case of error from
add_extent_to_dedupe().

Also, don't try dedupe if we ever only got one extent to succesfully be
added.

This patch also adds some asserts to help catch loop errors in the future.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>