xxhash: use xxhash directly
Don't sum up the checksums as the resulting digest will be different for the
same input depending on whether we use the one-shot or running checksum
functions.

This does result in a smaller digest, padded to the end with zeros which
doesn't look great (we can always fix that) but is correct.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>