csum-murmur3: shrink rem_buffer
We empty it at the top of add_to_running_checksum() if it would otherwise
become as large or larger than a 16 byte block. In that case, we know that
it can never be larger than 15 bytes. Change the definition to reflect this
and update the assert at the bottom of add_to_running_checksum().

Signed-off-by: Mark Fasheh <mfasheh@suse.de>