csum-murmur3: don't memset rem_buffer
This is overflowing the buffer and is not necessary anyway since we calloc
the struct which rem_buffer is a member of.

Since we define a constant for its size, we can use that to simplify
some of our asserts. This also removes a redundant assert from the top of
start_running_checksum().

Signed-off-by: Mark Fasheh <mfasheh@suse.de>