Use correct fiemap start in filerec_count_shared().
If we hit a file with more extents than our buffer can hold, we'll loop back
to do another fiemap on the file.  The start value wasn't being correctly
set, so that code was just looking up the same range over and over again.
Fix it so when we loop around fiemap is given the correct start offset.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>