remove_extent() might delete the 'dext' parent structure if our call makes the extent list empty. The code in clean_deduped() and dedupe_extent_list() is not taking this into account, resulting in some double frees.
Fix this by checking the return code of remove_extent(). If it returns
0, we know there are no longer any extents left on the structure and
we can discard it.

The thread pool code in dedupe_results() references the dext structure
after passing it to the thread pool. Since we are now aware that they
can be destroyed by the thread, update dedupe_results() to get the
next node pointer before the call to our thread pool. This avoids any
possibility of referencing a freed pointer. We add some locking around
the call here too in case the rbtree code might race with itself.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>
Git-commit: 31d0e35