Move the call to clean_deduped() back into dedupe_extent_list()
We need to call this after add_shared_extents() so that a physical offset
will have been stored on the extent structure. Otherwise, we're comparing a
bunch of initialized-to-zero offsets, which causes clean_deduped() to throw
out all the extents (thus killing any chance of dedupe).

I make a special return code for dedupe_extent_list in this case so that
our worker can catch it and return a non-error value.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>
Git-commit: fb15260c30c7671bff8be5548d18eb41d33b9e20