Retrieve file size from hashfile
We were storing file size correctly during the file scan phase but not
loading it up later when we re-read the hashfile. This didn't show up on
normal runs since we don't remove the filerec cache between steps, but would
show up immediately if we loaded filerecs from the hashfile directly.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>