make dedupe_ctxt blocksize 64 bits
Thanks to github user juliantaylor for pointing this out -
set_aligned_same_length() uses this variable to mask out the lower bits of
an unaligned length. If the blocksize variable is 32 bits though and the
length is 64, we'll get the top bits of our blocksize mask as zeros, which
gives us wildly incorrect answers.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>