Clean up prints during the dedupe phase
Print the hash id of dupe extents in the dupe table, and also during the
dedupe phase. This is intended to make it easier for users to follow the
status of a series of extents throughout the course of a run.

The prints during dedupe_extent_list also get cleaned up, and if verbose is
asked for, we give a print of each individual extent as it gets queued into
the dedupe context. All verbosity levels now get an error output if an
individual extent was unable to be deduped.

Lastly, I added a console mutex to avoid some of these prints spilling over
into each other. I've only had the issue with complete lines so we'll start
by only using the mutex when printing more than one line.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>
Git-commit: 6a9b5d5