Check if a dedupe request might fail
Bail out on a dedupe if we are on a kernel which requires an aligned length
and we have fewer than blocksize bytes remaining. Previously, we just let
the request go through and have the kernel sort it out. This is a valid
approach, but by short-circuiting here we avoid bubbling up the harmless
EINVAL that results when this happens.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>