Don't print some -EINVAL error in process_dedupe_results
We can get at least one case (file deduping itself) which can give us -EINVAL
on kernels before the 4.2 release. The best fix would be to not get them on
the list in the first place but this is sufficient at the moment. Eventually
we want to add more support for file self-dedupe anyway.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>