Only dedupe files whose mtime has changed
This involves a lot of moving parts. We make hashfiles reusable, allowing
the user to re-run duperemove with the same hashes from a previous run.

As a result, we have to manage the information in our db instead of just
dumping data as we do now. In particular, our list of dedupe targets now
comes from the database as well as the command line. In order to avoid extra
stats, we handle the command line files first, then walk the files in our
database. It's at this time that we can compare mtime in our filerecs
agains that of the database.

Lastly, at the find-dupes and dedupe stages, we avoid comparing/deduping
against files that haven't changed, with the exception that we'll pick at
least one target so there's always the ability to dedupe an extent.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>