Block dedupe to skip extent finding stage
We have a find-extents stage which is intended to collate duplicate blocks
into extent lists. The idea is to reduce the total number of dedupe calls,
and hopefully cut down on fragmentation.

On workloads with large hash buckets this algorithm can sometimes take a
long time. Block dedupe skips that stage and goes directly to the dedupe
portion. We're still deduping at pretty large blocks, and the target code in
dedupe_extent_list() will ensure requests go to the same block. So while we
won't get things lined up as perfectly, it should (in thoery) still go ok.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>