cleanup fiemap_iter_get_flags()
This was way to hacky before. Get rid of the gotos - now we cleanly iterate
over our fiemap context in a loop while looking for an appropriate extent.

We also handle the case where we close the ctxt but still have flags to
share by storing them on the fiemap context and retrieving them when needed.

Lastly add a small unit test to let us run this against a file without the
actual csuming that duperemove does.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>