Update handling of one-file-system
Firstly, we can make '-x' work on btrfs by tracking btrfs fsid in place of
device. This adds an extra statfs, so we move the check for one-file-system
inside of the check for a directory. This makes sense _anyway_ since we
only mount on directories.

Secondly, this means we can now turn on -x (one filesystem) by default. If
we had the fix described above we could have done this earlier. We also give
-x an optional argument so people can turn it off. -x being on by default
makes sense for duperemove - we can't dedupe between filesystems and the
file scan stage is heavy. In addition, reusable hashfiles make dealing with
crossing fs boundaries even more difficult.

Lastly, I added a small unit test to btrfs-util.c so that we can
independently check that the fsid manipulation works.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>