Remove '-x no' option
On second thought, don't allow turning of this off. Cross filesystem dedupe
does not exist and trying to make allowances for multiple fs inodes in
hashfiles is a lot of work for no gain. In particular we don't want to get
into situations where we have the same ino/subvol pair for different files
in the hash table.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>