Track when we run dedupe
We were using FILEREC_RESCANNED in dedupe to tell us whether a file could be
skipped. This is problematic however - FILEREC_RESCANNED only told us if the
file was rescanned during our current run. As a result files that were added
to the hashfile prior to a run with the dedupe flag set were being ignored
as candidates for dedupe.

We don't want to go about writing flags after every extent dedupe so use a
global sequence (dedupe_seq) instead. dedupe_seq starts at 0 and is
incremented only when we run with dedupe. On scan/rescan, files get a
sequence of dedupe_seq + 1. The dupe finding code then can simply ask
whether file->dedupe_seq <= dedupe_seq to see if a file can be skipped.

This patch completely replaces the FILEREC_RESCANNED flag.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>