Turn off memory journal
Previous change to bundle our file/hash updates into a transaction won't
work correctly unless we have a disk journal. Performance-wise, this doesn't
seem to make a noticable difference. We already disable syncing via pragma
so the behavior makes sense - the journal will be in the fs cache most of
the time so we're still accessing it at memory speeds.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>