Remove db record if a file was renamed
I thought we had a nice optimization but the files table is indexed by
filename, so it when we go to replaced the old record we run the risk of
replacing the wrong one. We could get around this with some fancy logic and
a bit of memory but that can wait for another day.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>