Set file->dedupe_seq if we drop out in add_file_db()
This shouldn't actually cause any problems as-is because we're marking it
for rescan (and rescan will write a new sequence). That said, we want to be
consistent with how we treat the filerec in the rest of our function, in
case we ever change the way csum stage sets dedupe_seq.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>