debug: calculate stats on find_all_dupes() coverage
In order to further hack on find_all_dupes() and remove_overlapping_extents(),
we need to be able to see how much 'coverage' of duplicated space a given set
of changes gives us. The most natural way to do this is per-file. So for each
file we compare the sum of our duplicated blocks (hash-tree blocks) with the
sum of our dedupe extents.

The data is printed in csv, there's also a block and extent dump on verbose
for immediate comparison. This is all compile-time as it costs cpu and for
now, is only interesting to anyone hacking on duperemove.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>