Improve coverage of dedupe results
remove_overlapping_extents() uses extent 'score' to determine which extents to
remove. It's also rather brutal with the removals - often removal of an
extent will trigger removal of an extent from another file.

Both of these factors give it somewhat unpredictable behavior and we lose
potential dedupe in many cases. Users have complained before that duperemove
can sometimes miss obvious points of dedupe.

Instead, we only consider for removal those extents which are entirely
encompassed within another extent. In addition, we don't remove any extents
that would result in another file losing dedupe.

This provides a lot more consistency - on my tests the new code tends to
tends to get ~100% coverage on a per file basis, whereas the old code swung
between ~100% and ~50% coverage. To be fair, 50% is the pathological case -
master is very close on average to giving ~100% coverage on most files.

The 'manyzeros' hash in duperemove-tests is a case where this change shows a
very noticable difference. The following example is just one file, but most
follow this pattern.

<filename>, <dupe bytes found>, <dupe bytes to submit>
Before:
/btrfs/testfile.0.01, 8589934592, 3534880768
After:
/btrfs/testfile.0.01, 8589934592, 8589934592

What makes this file set unique is that many of the files are copies of each
other. The score calculations find tons of small extents whereas this change
gives us a set of whole-file extents for dedupe.

Performance-wise this makes no discernable difference. In the cases where
we're catching more duplicate extents, there will be more dedupe work to do
but that's a logical and reasonable reason for a run to take longer than
before.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>