Revert 9e90d81 find_dupes: find all file compare possibilites 1st
This had to be done by hand, we add back in the old filerec compare tree at
the same time.

The memory usage of our compare tree in that patch turned out to be pretty
bad when we had a large # of files which shared duplicate blocks. The old
code might only be slightly better wrt memory, but this is in a pretty
constrained stage of the process already so every bit helps.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>