Make block-dedupe the default
The idea of find-dupes is a great one - we want to cut down on the number of
extent references placed on disk by building exents out of our dupe blocks
tree.

The problem is that we've never been able to get this to perform reasonably
well and give good dedupe results at the same time. The design doc in our
wiki has the full details but the most relevant excerpt would be:

We're trying to balance at least 3 very important resources:
 - cpu usage
 - memory usage
 - quality of dedupe

Right now we catch all possible extents (100% dedupe quality) at the expense
of a ton of memory and CPU. Turning down the quality in favor of fewer
expended resources tends to get us in situations where the pattern of dedupe
is seemingly random, or we always miss at least some obvious cases (such as
identical files).

We can continue to experiment until we get something that works well -
there's still many options going forward. In the meantime however, the
number of bug reports I have recieved where find-dupes is a severe
performance problem is too high. We want to ensure a smooth user experience,
especially for those with large dedupe sets so make find-dupes optional.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>