Add a Dockerfile for easier consumption of duperemove
This adds a Dockerfile to be able to use duperemove on platforms where it is
not natively available (e.g. Synology NAS systems).

Having a Docker image makes it easier to use duperemove. The user just needs
to bind-mount the directory he wants to dedupe into the container:

```
docker run -v /folder/to/dedupe:/to/dedupe duperemove -rdhv /to/dedupe
```

Signed-off-by: Jannick Fahlbusch <git@jf-projects.de>