Revert commit ac32d43 - "Make block-dedupe the default"
This is resulting in too much meta data fragmentation on the users end.
This sometimes pushes us into negative space savings - definitely not something
users expect from their dedupe tool. So revert the default back to our
extent search algorithm. Block dedupe is still there in case there is a use
for it.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>