Add helper functions to load file extents
The fiemap extent information we gather during our file scan can be
effectively used to limit or inform our extent matching code.

We add two helpers, one which loads a single arbitrary extent, and another
which loads all extents which do not have a duplicate extent in the extent
table.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>