We checksum every block of data twice during file scan - once for the block hash and once for the extent hash (which is a running hash over the entire extent).
We can get away with just adding each block hash into our running checksum,
so now we checksum every block plus an additional 8 bytes of data. This
should save us a decent amount of cpu.

We record this choice in the hashfile in case we want to go back to hashing
data for the extent hash.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>