Remove a bogus abort() in compare_extents()
We can get block->b_parent != orig->b_parent there if the fast forward loop
goes past extent_end and exits. The loop does check for this though so the
abort doesn't help catch any logic errors with it. We do have to catch this
condition though, but after the search and FF loops.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>