Rewrite cput core count/ht detection
Current code relies on the presence of "ht" flag in /proc/cpuinfo and
works only on Intel cpus. This is broken due to:

a) AMD cpu's also reporting HT flag
b) The flag reports whether the cpu's cpuid instruction knows about
HT, not if it's on or off.

Fix this by parsing
/sys/devices/system/cpu/cpuXX/topology/thread_siblings_list
directory file for every core. That way physical cores can be derived
based on the presence of thread_siblings_list and logical ones based
on the number of siblings for every core.