Always try to check for zerored buffered in ssize_t chunks
Currently is_block_zeroed will only utilize the optimized variant if
passed in buf_size  is aligned to sizeof(ssize_t). That's not really
necessary, the optimized code can be used for any buf_size so long as
it's at least sizeof(ssize_t). So rework the function so utilize for the
portion of the passed in buffer which is multiple of the size of
ssize_t. For the rest utilize the same byte comparison code.