Fix return value of csum_blocks
In case add_block_hash returns an error we simply break and return 0,
instead the error value (ENOMEM) should be propagated. Also change the
return value of add_block_hash to a negative so that when it's forwarded
between in add_block_hash->csum_blocks->csum_extent->csum_by_block it
can be correctly handled.