Properly fix truncated file handling
While trying to fix truncated file handling in fd2011333c1a ("Fix
handling of truncated files when scanning in block mode") I inadvertenly
broke file hashing. The end effect was that the last extent of a file
got only its first blocksize bytes hashed.

This commit fixes this and properly fixes truncated file handling. This
is achieved by setting the file size (which is the loop termination
condition) to the real file end when we encounter the last extent. This
ensures the whole range of valid bytes is iterated and that termination
would eventually occur.