fix handling of extents larger than 2 gigabytes
Internally the length of extents was managed in 4 byte quantities
despite the fiemap interface returning an 8 byte quantity. This open the
opportunity to wrap around in case we have an extent larger than 4
gigabytes. What's more the sqlite interface used the sqlite3_bind_int
function which further opened up the possibility of truncating the the
passed quantity, this effectively made the maximum size of extents
written to the database be 2 gigabytes. Larger ones were written as a
negative value.

Fix this by changing all internal variables dealing with values returned
from fiemap to utilize 64 bit quantities as well as switch to using
sqlite3_bind_int64, this still opens the possibility of truncating the
uint64 value returned from the kernel if there is an extent larger
than 9,223,372,036,854,775,807 bytes, but this borders on madness.

Fixes #275