duperemove: implement batching
We used to run a three-phases algorithm:
- list files
- hash files
- dedup files

When batching is enabled, phases two and three are interwined:
The hash phase hashes up to N files, and the run the dedup phase,
then hash up to N other files, and run the dedup phase etc.

This adds some overhead, because the thread pool is destroyed
for each hash phase. This also means that we have a head-of-line blocking,
this increases the elapsed time from a user perspective.

However, the dedup phase is only run against the N files.
Hence, both the results_tree and the hash_tree are kept small,
this greatly reduces the memory consumption.

This makes duperemove usable on very large dataset, or when partial
lookup is enabled.

Signed-off-by: Alexandre Bruyelles <git@jack.fr.eu.org>