duperemove: do not store block hashes unless we actually use them
A large part of the hashfile is consumed by block hashes.
It also consumes memory, because csum_whole_file keeps
all the file's block hashes in memory and then commit them to the hashfile.

When the user is doing full-extents deduplication, using the v3 hashfile
and without partial extents lookup, those block hashes are never used.

In such cases, this commit simply discards the hashes.

It shall be noted that options should be consistent over time: a file scanned
without the block will not be rescanned if, say, the user wants to do
partial extents lookup. In such situation, the user should perform a full scan
by deleting the hashfile.

Signed-off-by: Alexandre Bruyelles <git@jack.fr.eu.org>