file_scan: fix duplicate files
__add_file will detect if a file is hardlinked with another file.
In such cases, the file is simply ignored.
The code used to work with a filerec lookup. However, when we use the
norescan_files option, some files exist in the database but are never
loaded as filerecs, so the detection fails.

This commit replaces the filerec lookup by a hashfile lookup.
dbfile_load_one_filerec() will now gracefully handle missing entries.
All the remaining users of that function fetched their data from the hashfile,
so this should not be an issue.

Signed-off-by: Alexandre Bruyelles <git@jack.fr.eu.org>