file_scan: add_file_db: rescan the file when its size changed
In most cases, a changed size without a changed mtime means a fallocate
on the file.
However, if the filesystem is mounted with the noatime option, then mtime
is worthless and filesize is the only way to detect a change.

I believe the later case is more important than the former, because
a file modified with fallocate will probably have its content modified too.

Signed-off-by: Alexandre Bruyelles <git@jack.fr.eu.org>