file_scan: reimplement file scan
We used to:
- build a list of struct filerec that needs scan
- iterate over the list and process each file

This allowed us to (re)scan files from the database that was not
found on the arglist. On the other hand, it was memory-intensive:
you had to consume memory for every files ever seen.

This commit reworks the scan phase:
- the main thread create a worker pool
- it then lists the files, checks with the database if the file needs (re)scan,
  and pushes an item against the thread pool

Main benefits are:
- memory is freed as soon as a file is scanned
- if only a couple of files are passed in the arglist, there is few work to do

Main drawbacks are:
- files from database that are not on the arglist are not rescanned
- files that were scanned long ago and were since delete are not removed
  from the hashfile.

Because of that, the 'rescan_files' option, added in v0.13 is removed, its
behavior is hardcoded.

Signed-off-by: Alexandre Bruyelles <git@jack.fr.eu.org>