file_scan: lock the targeted filesystem based on its uuid
Since forever, we used to use btrfs fsid to detect if
a given directory would cross to another filesystem.
For non-btrfs, we used its dev_t.

For btrfs, we stored the file subvolume id in the hashfile.
For non-btrfs, we stored the dev_t inplace.

Indeed:
- subvolume id is not a unique ID. For instance, all top-level
  subvolumes has an id 5.
- dev_t is not persistent and may change if you umount and remount
  the filesystem.

This commit tries to improve that situation:
- the filesystem's UUID is grabbed and stored in the hashfile
- every path passed by the user and every directory is checked against
  that uuid to see if it is not part of another filesystem

To quicken the operation on non-btrfs filesystem, the current dev_t is
fetched on the first valid path we encounter. All futures paths can
then check the dev_t and avoid the (somewhat costly) uuid fetch.

Fetching UUID on btrfs is available through an ioctl.
However, for non-brtfs, the task is not as simple as I would like.
For those situations, libblkid is used. If you are not using duperemove as
root, it requires a prefilled cache, which can be created using:
	sudo blkid

Signed-off-by: Alexandre Bruyelles <git@jack.fr.eu.org>