file_scan: do not invalidate a file renamed since the last run
Path are stored in the hashfile because we need to open(2) it for
deduplication.
However, during the scan file, path is not really used: we identify
the files based on their inode number (and subvolume id for btrfs).

Between runs, a file may be renamed: in this case, its ino won't change,
nor its content, so the scan phase used to do nothing.
Yet later, the dedupe would try to open the old path, fails and
conclude that the file were removed.

This patch improve the scan phase to detect the rename and update
the hashfile accordingly.

Signed-off-by: Alexandre Bruyelles <git@jack.fr.eu.org>