dbfile: fix whole file duplicates selection
The previous query only returned files if duplicates
were found in the same batch iteration.

As a result, if the first occurence appeared in the batch n
and the second appeared in the batch n+1, the query would
return nothing, and nothing would be deduped.

This commit tries to fix this behavior, the query should
return any duplicates, except future ones, and only if
the duplicates is somehow related to the current batch set.

Fixes #347

Signed-off-by: Alexandre Bruyelles <git@jack.fr.eu.org>